extracted window properties from apps
